% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a CSV file as a tibble.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the CSV file}
}
\value{
The function returns the contents of the CSV file as a tibble.
}
\description{
The function is a simple wrapper around readr::read_csv. An error is raised if the file does not exits.
}
\note{
tbl_df is deprecated, use tibbble::as_tibble instead
}
\examples{
fars_read("data/accident_2013.csv.bz2")

}
