% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads a FARS file, identified just by its year, as a tibble.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years, or a single year. The function will try to find the correct
FARS file for each year in the current directory.}
}
\value{
The function returns the contents of the FARS files as a list of tibbles, one per year,
 with just month and year per accident, for all the years in the given years list. If a year has no
 data file, a warning is shown and NULL is returned.
}
\description{
Reads a FARS file, identified just by its year, as a tibble.
}
\examples{
fars_read_years(c(2013, 2014))
fars_read_years(2013)

}
