% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads and summarize FARS data for the given years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years, or a single year. The function will try to find the correct
FARS file for each year in the current directory.}
}
\value{
The function returns the summary of the FARS files for the given years,
 as number of accidents per month and year.
}
\description{
FARS files should be on the working directory.
}
\examples{
fars_summarize_years(c(2013, 2014))
fars_summarize_years(2013)

}
