% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates correct FARS file names}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year, the function will try to convert it to an integer.}
}
\value{
The function returns the file name of the CSV FARS file for the given year. No path is added.
}
\description{
A function helper for the creation of correct file names for the
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System data,
}
\details{
This function takes a year (any type with coversion to numeric)
and spells out the filename (no path) for the FARS file of that year.
}
\examples{
make_filename(2013)
make_filename("2013")

}
